# api/openapi.yaml
openapi: 3.0.3
info:
  title: VimeSrv API
  description: REST API for VimeSrv - Home network video media server with DASH/HLS transcoding
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API token obtained from /api/v1/auth/login
    StreamAuth:
      type: apiKey
      in: query
      name: token
      description: Short-lived stream token obtained from /api/v1/auth/stream-token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_REQUEST"
            message:
              type: string
              example: "Invalid request body"
            details:
              type: object
...
    MovieSummary:
      type: object
      properties:
        media_id:
          type: string
        duration:
          type: integer
        resolution:
          type: string
        status:
          type: string
        enrichment_status:
          type: string
        created_at:
          type: string
        transcode_status:
          type: string
          enum: [none, pending, completed]
        movie_metadata_id:
          type: integer
          format: int64
        title:
          type: string
        year:
          type: string
        poster_path:
          type: string
        backdrop_path:
          type: string
        vote_average:
          type: number
        genres:
          type: string
    SeriesSummary:
      type: object
      properties:
        series_metadata_id:
          type: integer
          format: int64
        tmdb_id:
          type: integer
        name:
          type: string
        year:
          type: string
        poster_path:
          type: string
        backdrop_path:
          type: string
        vote_average:
          type: number
        genres:
          type: string
        number_of_seasons:
          type: integer
        number_of_episodes:
          type: integer
        available_episodes:
          type: integer
    EpisodeSummary:
      type: object
      properties:
        media_id:
          type: string
        duration:
          type: integer
        status:
          type: string
        transcode_status:
          type: string
        episode_metadata_id:
          type: integer
          format: int64
        season_number:
          type: integer
        episode_number:
          type: integer
        name:
          type: string
        overview:
          type: string
        air_date:
          type: string
        still_path:
          type: string
        vote_average:
          type: number
    SeasonSummary:
      type: object
      properties:
        season_metadata_id:
          type: integer
          format: int64
        season_number:
          type: integer
        name:
          type: string
        overview:
          type: string
        poster_path:
          type: string
        air_date:
          type: string
        episode_count:
          type: integer
        episodes:
          type: array
          items:
            $ref: '#/components/schemas/EpisodeSummary'
    SeriesDetail:
      allOf:
        - $ref: '#/components/schemas/SeriesSummary'
        - type: object
          properties:
            overview:
              type: string
            seasons:
              type: array
              items:
                $ref: '#/components/schemas/SeasonSummary'
            similar_series:
              type: array
              items:
                $ref: '#/components/schemas/SimilarSeriesItem'
    SimilarMovieItem:
      type: object
      properties:
        tmdb_id:
          type: integer
        title:
          type: string
        poster_path:
          type: string
        release_date:
          type: string
        year:
          type: string
        vote_average:
          type: number
        in_library:
          type: boolean
        media_id:
          type: string
          description: Set if in_library is true
    SimilarSeriesItem:
      type: object
      properties:
        tmdb_id:
          type: integer
        name:
          type: string
        poster_path:
          type: string
        first_air_date:
          type: string
        year:
          type: string
        vote_average:
          type: number
        in_library:
          type: boolean
        series_metadata_id:
          type: integer
          format: int64
          description: Set if in_library is true
    CollectionMovieItem:
      type: object
      properties:
        tmdb_id:
          type: integer
        title:
          type: string
        poster_path:
          type: string
        release_date:
          type: string
        year:
          type: string
        vote_average:
          type: number
        in_library:
          type: boolean
        media_id:
          type: string
          description: Set if in_library is true
        is_current:
          type: boolean
          description: True if this is the movie being viewed
    MovieCollectionInfo:
      type: object
      properties:
        collection_id:
          type: integer
        name:
          type: string
        overview:
          type: string
        poster_path:
          type: string
        backdrop_path:
          type: string
        movies:
          type: array
          items:
            $ref: '#/components/schemas/CollectionMovieItem'
        position:
          type: integer
          description: 1-based position of current movie in collection
        total_movies:
          type: integer
    CreditPerson:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tmdb_person_id:
          type: integer
        name:
          type: string
        character:
          type: string
        job:
          type: string
        profile_path:
          type: string
    MovieDetail:
      allOf:
        - $ref: '#/components/schemas/MovieSummary'
        - type: object
          properties:
            original_title:
              type: string
            tagline:
              type: string
            overview:
              type: string
            release_date:
              type: string
            runtime:
              type: integer
            movie_status:
              type: string
            imdb_id:
              type: string
            tmdb_id:
              type: integer
            certification:
              type: string
            cast:
              type: array
              items:
                $ref: '#/components/schemas/CreditPerson'
            directors:
              type: array
              items:
                $ref: '#/components/schemas/CreditPerson'
            crew:
              type: array
              items:
                $ref: '#/components/schemas/CreditPerson'
            similar_movies:
              type: array
              items:
                $ref: '#/components/schemas/SimilarMovieItem'
            collection:
              $ref: '#/components/schemas/MovieCollectionInfo'
    MovieDetailResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/MovieDetail'
    UnmatchedMediaSummary:
      type: object
      properties:
        media_id:
          type: string
        filename:
          type: string
        title:
          type: string
        duration:
          type: integer
        resolution:
          type: string
        enrichment_status:
          type: string
        created_at:
          type: string
    MoviesListResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/MovieSummary'
            total:
              type: integer
            page:
              type: integer
            per_page:
              type: integer
    SeriesListResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/SeriesSummary'
    SeriesDetailResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/SeriesDetail'
    RecentListResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/MovieSummary'
    UnmatchedListResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/UnmatchedMediaSummary'
            total:
              type: integer
            page:
              type: integer
            per_page:
              type: integer
paths:
...
  /api/v1/library/movies:
    get:
      tags: [Library]
      summary: List movies
      description: Returns a paginated list of movies with metadata
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: lang
          in: query
          description: Language for metadata translations
          schema:
            type: string
            default: en
      responses:
        '200':
          description: Movies list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviesListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/library/movies/{id}:
    get:
      tags: [Library]
      summary: Get movie details
      description: Returns detailed information about a specific movie with metadata
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Media file ID
          schema:
            type: string
        - name: lang
          in: query
          description: Language for metadata translations
          schema:
            type: string
            default: en
      responses:
        '200':
          description: Movie details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailResponse'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/library/series:
    get:
      tags: [Library]
      summary: List series
      description: Returns a list of series with metadata
      security:
        - BearerAuth: []
      parameters:
        - name: lang
          in: query
          description: Language for metadata translations
          schema:
            type: string
            default: en
        - name: include_empty
          in: query
          description: Include series with no available episodes
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Series list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeriesListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/library/series/{id}:
    get:
      tags: [Library]
      summary: Get series details
      description: Returns detailed information about a series with all seasons and episodes
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Series metadata ID
          schema:
            type: integer
            format: int64
        - name: lang
          in: query
          description: Language for metadata translations
          schema:
            type: string
            default: en
      responses:
        '200':
          description: Series details with seasons and episodes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeriesDetailResponse'
        '400':
          description: Invalid series ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Series not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/library/recent:
    get:
      tags: [Library]
      summary: List recently added
      description: Returns recently added movies and episodes
      security:
        - BearerAuth: []
      parameters:
        - name: lang
          in: query
          description: Language for metadata translations
          schema:
            type: string
            default: en
      responses:
        '200':
          description: Recently added media
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/library/unmatched:
    get:
      tags: [Library]
      summary: List unmatched media
      description: Returns media files without linked metadata
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Unmatched media list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnmatchedListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...